requirements.txt
streamlit
langchain
langgraph
langchain_core
langchain_community
openai
loguru
langchain-groq
fastapi
uvicorn
python-dotenv
numpy<2.0.0
SpeechRecognition
pyttsx3
vosk
sounddevice
PyAudio
apscheduler
python-jose[cryptography]
emails
jinja2